Add setup for local xmodmap
This setup now uses a global `xmodmap/init` and a local `xmodmap/init.local`. This way we can do common things like swap `Caps_Lock` and `Control_L` in the global setup, and do things that depend on the machine like swap `mod1` (alt) and `mod4` (super) in a local setup that does not get checked in to source contol.